backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
motor==3.3.2
qrcode==7.4.2
pillow==10.1.0
python-dotenv==1.0.0
joblib==1.3.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
schedule==1.2.0
pywin32==306 (Windows only)
pyobjc-core (macOS only)